Agbo

+++ src/redux/Xetima/User/UserActions.js
@@ -7,10 +7,11 @@ import{
     SELECT_ONE_USER_FAIL,
     DELETE_USER,
     DELETE_USER_SUCCESS,
-    DELETE_USER_FAIL
+    DELETE_USER_FAIL,
+    RESET_USERS_STATE
 } from './UserTypes'
 
-import { BACKEND_BASE_URL, headerIncluder } from "../../../common_variables";
+import { BACKEND_BASE_URL, headerIncluder, NO_OF_TRIAL_COUNTER } from "../../../common_variables";
 import * as Validator from 'validatorjs';
 import validateModule from "../../../validation/validate_module";
 import { postRequest, getRequest } from "../../axios_call";
@@ -19,7 +20,7 @@ import {CHANGE_USER_OBJECT} from "../Login/LoginActionTypes";
 const getAllUsersAction = () => {
     return {
         type: GET_ALL_USERS,
-        message:''
+        message:'Loading...'
     };
 };
 
@@ -38,29 +39,47 @@ const getAllUsersActionFailure = (message) => {
     }
 }
 
-export const getUsersAction = (loginData) => async (dispatch) => {
+export const getUsersAction = (loginData, allUsers, counter = 0) => {
 
-    dispatch(getAllUsersAction());
+    return async (dispatch) => {
+
     try{
+        counter++;
+        dispatch(getAllUsersAction());
+
         if(loginData.isLogged === true){
-            let handleagetUsersAction = await getRequest(BACKEND_BASE_URL+"users/all_users/user", headerIncluder(loginData.user_data.token) );
-            let returnedObject = handleagetUsersAction.data;
-            // console.log(handleagetUsersAction)
-            let {status, message, data} = returnedObject;
-            let {all_users} = data;
-            if(status === true){
-                dispatch({type:GET_ALL_USERS_SUCCESS, payload:all_users});
+
+            if(allUsers.length > 0){
+                dispatch(getAllUsersActionSuccess({data:allUsers, message:'' }));
             }else{
-                validateModule.handleErrorStatement(message, '', 'on', 'no', 'no');
-                dispatch({
-                    type:GET_ALL_USERS_FAIL,
-                    message:message
-                });
+                let handleagetUsersAction = await getRequest(`${BACKEND_BASE_URL}users/all_users/user`, headerIncluder(loginData.user_data.token) );
+                let returnedObject = handleagetUsersAction.data;
+                // console.log(handleagetUsersAction)
+                let {status, message, data} = returnedObject;
+                let {all_users} = data;
+                if(status === true){
+                    dispatch(getAllUsersActionSuccess({data:all_users, message:'' }));
+                }else{
+                    validateModule.handleErrorStatement(message, '', 'on', 'no', 'no');
+                    dispatch({
+                        type:GET_ALL_USERS_FAIL,
+                        message:message
+                    });
+                }
             }
         }
     }catch(err){
+
+        /*if(err.message === 'Network Error'){
+            if(counter < NO_OF_TRIAL_COUNTER){
+                dispatch(getAllUsersActionFailure(err.message));
+                return getUsersAction(loginData, allUsers, counter);
+            }
+        }*/
+        validateModule.handleErrorStatement({general_error:[err.message]}, '', 'on', 'no', 'no');
         dispatch(getAllUsersActionFailure(err.message));
     }
+}
 }
 
 //............................................ Get single user ..........................................................//
@@ -115,7 +134,7 @@ export const selectOneUserAction = (loginData, unique_id) => async (dispatch) =>
 const deleteUserAction = () => {
     return {
         type: DELETE_USER,
-        message:''
+        message:'Loading....'
     };
 };
 
@@ -134,28 +153,38 @@ const deleteUserActionFailure = (message) => {
     }
 }
 
-export const deleteUsersAction = (loginData, unique_id, type_of_user) => async (dispatch) => {
+export const deleteUsersAction = ({unique_id, type_of_user, loginData}) => async (dispatch) => {
+    let reVal = window.confirm('Do you really want to delete user ?');
+    if(reVal === true){
+        dispatch(deleteUserAction());
+        try{
 
-    dispatch(deleteUserAction());
-    try{
-        if(loginData.isLogged === true){
-            let handleDeleteUserAction = await getRequest(BACKEND_BASE_URL+`users/delete_user/${unique_id}/${type_of_user}`, headerIncluder(loginData.user_data.token) );
-            let returnedObject = handleDeleteUserAction.data;
-            console.log(handleDeleteUserAction, "deleted")
-            alert(handleDeleteUserAction)
-            let {status, message, data} = returnedObject;
-            // let {all_users} = data;
-            if(status === true){
-                dispatch(deleteUserActionSuccess(message, data));
-            }else{
-                validateModule.handleErrorStatement(message, '', 'on', 'no', 'no');
-                dispatch({
-                    type:DELETE_USER_FAIL,
-                    message:message
-                });
+            if(loginData.isLogged === true){
+                let handleDeleteUserAction = await getRequest(`${BACKEND_BASE_URL}users/delete_user/${unique_id}/${type_of_user}`, headerIncluder(loginData.user_data.token) );
+                let returnedObject = handleDeleteUserAction.data;
+                let {status, message, message_type, data} = returnedObject;
+                let {all_users} = data;
+                if(status === true){
+                    dispatch(deleteUserActionSuccess(message, all_users));
+                }else{
+                    validateModule.handleErrorStatement(message, '', 'on', 'no', 'no');
+                    dispatch({
+                        type:DELETE_USER_FAIL,
+                        message:message
+                    });
+                }
             }
+        }catch(err){
+            dispatch(deleteUserActionFailure(err.message));
         }
-    }catch(err){
-        dispatch(deleteUserActionFailure(err.message));
     }
+}
+
+
+
+
+export const resetUserState = () => async (dispatch) => {
+    dispatch({
+        type:RESET_USERS_STATE,
+    });
 }
